pyproject.toml
[project]
name = "bank-project"
version = "0.1.0"
description = ""
authors = [
    {name = "Anna Novikova",email = "anna.novikova.dev@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "requests (>=2.32.3,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.lint.dependencies]
flake8 = "^7.2.0"
mypy = "^1.15.0"
black = "^25.1.0"
isort = "^6.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
pytest-cov = "^6.1.1"

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
exclude = 'venv'

[tool.black]
line-length = 119
exclude = '''
/\.git
'''

[tool.isort]
line_length = 119


